Break circular reference
